// src/api/__generated__/wallet/model/completeTaskResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama User Wallet API
 * 用户、钱包、任务中心接口
 * OpenAPI spec version: v1.0
 */
import type { CompleteTaskResponseStatus } from './completeTaskResponseStatus';
import type { CompleteTaskResponseType } from './completeTaskResponseType';

/**
 * 任务完成返回
 */
export interface CompleteTaskResponse {
  /** 任务 ID（后管分配，10–99） */
  taskId?: number;
  /** 任务类型 */
  type?: CompleteTaskResponseType;
  /** 本次入账积分 */
  points?: number;
  /** 任务状态 */
  status?: CompleteTaskResponseStatus;
}