// src/api/__generated__/wallet/model/taskStatusResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama User Wallet API
 * 用户、钱包、任务中心接口
 * OpenAPI spec version: v1.0
 */
import type { TaskStatusItem } from './taskStatusItem';

/**
 * 活动任务列表状态
 */
export interface TaskStatusResponse {
  /** 任务状态项列表 */
  tasks?: TaskStatusItem[];
}